+++ src/main/kotlin/com/markskelton/OneDarkThemeManager.kt
@@ -1,25 +1,26 @@
 package com.markskelton
 
 import com.intellij.ide.plugins.PluginManagerCore
-import com.intellij.ide.ui.LafManagerListener
-import com.intellij.ide.ui.laf.LafManagerImpl
-import com.intellij.ide.ui.laf.TempUIThemeBasedLookAndFeelInfo
-import com.intellij.ide.ui.laf.UIThemeBasedLookAndFeelInfo
 import com.intellij.openapi.application.ApplicationManager
-import com.intellij.openapi.editor.colors.EditorColorsListener
-import com.intellij.openapi.editor.colors.EditorColorsManager
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import com.intellij.util.messages.MessageBusConnection
 import com.markskelton.notification.Notifications
-import com.markskelton.settings.THEME_CONFIG_TOPIC
-import com.markskelton.settings.ThemeConfigListener
 import com.markskelton.settings.ThemeSettings
 import java.util.*
 
+enum class OneDarkThemes {
+  REGULAR, ITALIC, VIVID, VIVID_ITALIC
+}
+
 object OneDarkThemeManager {
   private lateinit var messageBus: MessageBusConnection
-  const val ONE_DARK_ID = "f92a0fa7-1a98-47cd-b5cb-78ff67e6f4f3"
+  val THEMES = mapOf(
+    "f92a0fa7-1a98-47cd-b5cb-78ff67e6f4f3" to OneDarkThemes.REGULAR,
+    "1a92aa6f-c2f1-4994-ae01-6a78e43eeb24" to OneDarkThemes.ITALIC,
+    "4b6007f7-b596-4ee2-96f9-968d3d3eb392" to OneDarkThemes.VIVID,
+    "4f556d32-83cb-4b8b-9932-c4eccc4ce3af" to OneDarkThemes.VIVID_ITALIC
+  )
   private const val PLUGIN_ID = "com.markskelton.one-dark-theme"
 
   fun registerStartup(project: Project) {
@@ -27,8 +28,6 @@ object OneDarkThemeManager {
       registerUser()
 
       attemptToDisplayUpdates()
-
-      applyConfigurableTheme()
 
       subscribeToEvents()
     }
@@ -44,7 +43,6 @@ object OneDarkThemeManager {
     getVersion().ifPresent { currentVersion ->
       if (ThemeSettings.instance.version != currentVersion) {
         ThemeSettings.instance.version = currentVersion
-        ThemeSettings.instance.customSchemeSet = false
         ApplicationManager.getApplication().invokeLater {
           Notifications.displayUpdateNotification(currentVersion)
         }
@@ -58,52 +56,5 @@ object OneDarkThemeManager {
 
   private fun subscribeToEvents() {
     messageBus = ApplicationManager.getApplication().messageBus.connect()
-    messageBus.subscribe(THEME_CONFIG_TOPIC, object : ThemeConfigListener {
-      override fun themeConfigUpdated(themeSettings: ThemeSettings) {
-        if (isCurrentTheme()) {
-          ThemeSettings.instance.customSchemeSet = false
-          LafManagerImpl.getInstance().setCurrentLookAndFeel(
-            ThemeConstructor.constructNewTheme(themeSettings)
-          )
-        }
-      }
-    })
-
-    messageBus.subscribe(LafManagerListener.TOPIC, LafManagerListener {
-      val currentLaf = it.currentLookAndFeel
-      if (currentLaf is UIThemeBasedLookAndFeelInfo) {
-        when {
-          currentLaf !is TempUIThemeBasedLookAndFeelInfo &&
-            isOneDarkTheme(currentLaf) -> setOneDarkTheme()
-        }
-      }
-    })
-
-    messageBus.subscribe(EditorColorsManager.TOPIC, EditorColorsListener {
-      ThemeSettings.instance.customSchemeSet =
-        it != null && !it.metaProperties.containsKey("oneDarkScheme")
-    })
   }
-
-  private fun isOneDarkTheme(uiThemeBasedLookAndFeelInfo: UIThemeBasedLookAndFeelInfo): Boolean =
-    uiThemeBasedLookAndFeelInfo.theme.id == ONE_DARK_ID
-
-  private fun applyConfigurableTheme() {
-    if (isCurrentTheme() && !ThemeSettings.instance.customSchemeSet) {
-      setOneDarkTheme()
-    }
-  }
-
-  private fun setOneDarkTheme() {
-    LafManagerImpl.getInstance().setCurrentLookAndFeel(
-      ThemeConstructor.useExistingTheme()
-    )
-  }
-
-  fun isCurrentTheme(): Boolean =
-    when (val currentLaf = LafManagerImpl.getInstance().currentLookAndFeel) {
-      is UIThemeBasedLookAndFeelInfo -> currentLaf.theme.id == ONE_DARK_ID
-      is TempUIThemeBasedLookAndFeelInfo -> currentLaf.theme.id == ONE_DARK_ID
-      else -> false
-    }
 }