+++ src/main/kotlin/com/markskelton/notification/Notifications.kt
@@ -2,20 +2,15 @@ package com.markskelton.notification
 
 import com.intellij.ide.plugins.PluginManagerCore.getPlugin
 import com.intellij.ide.plugins.PluginManagerCore.getPluginOrPlatformByClassName
-import com.intellij.notification.Notification
-import com.intellij.notification.NotificationAction
 import com.intellij.notification.NotificationDisplayType
 import com.intellij.notification.NotificationGroup
 import com.intellij.notification.NotificationListener
 import com.intellij.notification.NotificationType
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.options.ShowSettingsUtil
-import com.markskelton.settings.ThemeSettingsUI
 
 val UPDATE_MESSAGE: String = """
       What's New?<br>
       <ul>
-        <li>Theme is now configurable through settings.</li>
+        <li>Added Android Log Cat colors.</li>
       </ul>
       <br>Please see the <a href="https://github.com/one-dark/jetbrains-one-dark-theme/blob/master/CHANGELOG.md">Changelog</a> for more details.
       <br>
@@ -42,13 +37,6 @@ object Notifications {
       NotificationType.INFORMATION,
       NotificationListener.URL_OPENING_LISTENER
     )
-      .addAction(SettingsAction("Show Settings"))
       .notify(null)
-  }
-}
-
-class SettingsAction(text: String) : NotificationAction(text) {
-  override fun actionPerformed(e: AnActionEvent, notification: Notification) {
-    ShowSettingsUtil.getInstance().showSettingsDialog(e.project, ThemeSettingsUI.THEME_SETTINGS_DISPLAY_NAME)
   }
 }